Issue #151 CoreMark optimisation: change matrix array indexes from int to short. Saves about 264 ticks. (A: 2349->2086, B: 1728->1463)

+++ src/lib/bm_coremark/java/javax/rtcbench/CoreMatrix.java
@@ -42,7 +42,7 @@ public class CoreMatrix {
 	// C MATRES -> Java int
 
 	public static class MatParams {
-		public int N;
+		public short N;
 		public short[] A;
 		public short[] B;
 		public int[] C;
@@ -59,7 +59,7 @@ public class CoreMatrix {
 		changing the matrix values slightly by a constant amount each time.
 	*/
 	static short core_bench_matrix(MatParams p, short seed, short crc) {
-		int N=p.N;
+		short N=p.N;
 		int[] C=p.C;
 		short[] A=p.A;
 		short[] B=p.B;
@@ -94,7 +94,7 @@ public class CoreMatrix {
 
 		After the last step, matrix A is back to original contents.
 	*/
-	static short matrix_test(int N, int[] C, short[] A, short[] B, short val) {
+	static short matrix_test(short N, int[] C, short[] A, short[] B, short val) {
 		short crc=0;
 		short clipval=matrix_big(val);
 
@@ -180,7 +180,7 @@ public class CoreMatrix {
 		
 		Otherwise, reset the accumulator and add 10 to the result.
 	*/
-	static short matrix_sum(int N, int[] C, short clipval) {
+	static short matrix_sum(short N, int[] C, short clipval) {
 		int tmp=0,prev=0,cur=0;
 		short ret=0;
 		short i,j;
@@ -204,8 +204,8 @@ public class CoreMatrix {
 		Multiply a matrix by a constant.
 		This could be used as a scaler for instance.
 	*/
-	static void matrix_mul_const(int N, int[] C, short[] A, short val) {
-		int i,j;
+	static void matrix_mul_const(short N, int[] C, short[] A, short val) {
+		short i,j;
 		for (i=0; i<N; i++) {
 			for (j=0; j<N; j++) {
 				C[i*N+j]=(int)A[i*N+j] * (int)val;
@@ -216,8 +216,8 @@ public class CoreMatrix {
 	/* Function: matrix_add_const
 		Add a constant value to all elements of a matrix.
 	*/
-	static void matrix_add_const(int N, short[] A, short val) {
-		int i,j;
+	static void matrix_add_const(short N, short[] A, short val) {
+		short i,j;
 		for (i=0; i<N; i++) {
 			for (j=0; j<N; j++) {
 				A[i*N+j] += val;
@@ -229,8 +229,8 @@ public class CoreMatrix {
 		Multiply a matrix by a vector.
 		This is common in many simple filters (e.g. fir where a vector of coefficients is applied to the matrix.)
 	*/
-	static void matrix_mul_vect(int N, int[] C, short[] A, short[] B) {
-		int i,j;
+	static void matrix_mul_vect(short N, int[] C, short[] A, short[] B) {
+		short i,j;
 		for (i=0; i<N; i++) {
 			C[i]=0;
 			for (j=0; j<N; j++) {
@@ -243,8 +243,8 @@ public class CoreMatrix {
 		Multiply a matrix by a matrix.
 		Basic code is used in many algorithms, mostly with minor changes such as scaling.
 	*/
-	static void matrix_mul_matrix(int N, int[] C, short[] A, short[] B) {
-		int i,j,k;
+	static void matrix_mul_matrix(short N, int[] C, short[] A, short[] B) {
+		short i,j,k;
 		for (i=0; i<N; i++) {
 			for (j=0; j<N; j++) {
 				C[i*N+j]=0;
@@ -260,8 +260,8 @@ public class CoreMatrix {
 		Multiply a matrix by a matrix, and extract some bits from the result.
 		Basic code is used in many algorithms, mostly with minor changes such as scaling.
 	*/
-	static void matrix_mul_matrix_bitextract(int N, int[] C, short[] A, short[] B) {
-		int i,j,k;
+	static void matrix_mul_matrix_bitextract(short N, int[] C, short[] A, short[] B) {
+		short i,j,k;
 		for (i=0; i<N; i++) {
 			for (j=0; j<N; j++) {
 				C[i*N+j]=0;
